+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [unreleased]
+
+## [1.0.6-dev.8] - 2018-10-19
+
+- support other script languages (such as typescript) inside Vue files
+
 ## [1.0.6-dev.7] - 2018-10-09
 
 - utilize the cache mechanism of dependency-tree to cache resolved dependencies